Added How to Contribute To This Project section
In-person teams have sticky-notes on a whiteboard.  Remote teams have the github issue tracker.  This is the simplest way I can think of to create a computerized "todo, doing, done" board. 

(Trello is an alternative.  However, I think it is useful to keep the number of places my teams have to check for new work to a minimum.)